[Cache] optimize Cache implementation
Use code generation & memoization to generate functions specific to any
number of path entries. Also add a get/setter that can laziliy
initialize cache values.

Disclaimer: I modified only it's use in the Unit module, dunno if there
are more places.

On a completely unscientific benchmark doing

  local t = os.clock();
  for i = 1, 100000 do
    local path = "X."..i..".Y";
    local b = Cache.Get(path) or Cache.Set(path, i);
  end
  print(os.clock() - t);

for the current implementation and

  local t = os.clock();
  for i = 1, 100000 do
    local b = Cache.Get("X", i, "Y", function() return i end);
  end
  print(os.clock() - t);

for the new one reduces the final printed time from .832 to .073 which
is ~10 times faster.

On a side note, if we really care about cache performance even using
this implementation is suboptimal since this is still 2 times slower
than an unrolled implementation like

  local function GetXiY(i)
    local cX = Cache.X; if not cX then Cache.X = { [i] = { Y = i } }; return i; end
    local ci = cX[i]; if not ci then cX[i] = { Y = i }; return i; end
    ci.Y = i;
    return i;
  end

which takes only .037 for the same loop.

If we really care reaching this we'll have to rework both the Cache &
the Core addons to do code-gen basically everywhere (I don't have a
prototype for this yet but I believe I know how this can be done
somewhat elegantly).